refactor signChallenge
