+++ packages/app/features/auth/sign-in/sign-in-form.tsx
@@ -10,7 +10,7 @@ import {
   useToastController,
 } from '@my/ui'
 import { api } from 'app/utils/api'
-import { signChallenge } from 'app/utils/userop'
+import { signChallenge } from 'app/utils/signChallenge'
 import { useState } from 'react'
 import { useRouter } from 'solito/router'
 import { bytesToHex, hexToBytes } from 'viem'
