+++ packages/app/utils/userop.ts
@@ -1,4 +1,3 @@
-import { signWithPasskey } from '@daimo/expo-passkeys'
 import {
   baseMainnetBundlerClient,
   entryPointAddress,
@@ -20,9 +19,7 @@ import {
 import {
   bytesToHex,
   concat,
-  encodeAbiParameters,
   encodeFunctionData,
-  getAbiItem,
   getContract,
   hexToBytes,
   isHex,
@@ -35,12 +32,12 @@ import {
 import { useEstimateFeesPerGas } from 'wagmi'
 import { useQuery as useWagmiQuery, type UseQueryReturnType } from 'wagmi/query'
 import { assert } from './assert'
-import { parseAndNormalizeSig, parseSignResponse } from './passkeys'
 import { throwIf } from './throwIf'
 import { defaultUserOp } from './useUserOpTransferMutation'
 import { baseMainnetClient } from './viem'
 import { useMemo } from 'react'
 import { debug as debugBase } from './debug'
+import { signChallenge } from './signChallenge'
 
 const debug = debugBase('app:utils:userop')
 
@@ -126,70 +123,6 @@ export function generateChallenge({
     challenge,
     versionBytes,
     validUntilBytes,
-  }
-}
-
-/**
- * Signs a challenge using the user's passkey and returns the signature in a format that matches the ABI of a signature
- * struct for the SendVerifier contract.
- * @param challenge - The challenge to sign encoded as a 0x-prefixed hex string.
- * @param rawIdsB64 - The list of raw ids to use for signing. Required for Android and Chrome.
- * @returns The signature in a format that matches the ABI of a signature struct for the SendVerifier contract.
- */
-export async function signChallenge(
-  challenge: Hex,
-  allowedCredentials: { id: string; userHandle: string }[]
-) {
-  const challengeBytes = hexToBytes(challenge)
-  const challengeB64 = Buffer.from(challengeBytes).toString('base64')
-  const sign = await signWithPasskey({
-    domain: window.location.hostname,
-    challengeB64,
-    rawIdsB64: allowedCredentials.map(({ id }) => id), // pass the raw ids to the authenticator
-  })
-  // handle if a non-resident passkey is used so no userHandle is returned
-  sign.passkeyName =
-    sign.passkeyName ?? allowedCredentials.find(({ id }) => id === sign.id)?.userHandle ?? ''
-  assert(!!sign.passkeyName, 'No passkey name found')
-  const signResult = parseSignResponse(sign)
-  const clientDataJSON = signResult.clientDataJSON
-  const authenticatorData = bytesToHex(signResult.rawAuthenticatorData)
-  const challengeLocation = BigInt(clientDataJSON.indexOf('"challenge":'))
-  const responseTypeLocation = BigInt(clientDataJSON.indexOf('"type":'))
-  const { r, s } = parseAndNormalizeSig(signResult.derSig)
-  const webauthnSig = {
-    authenticatorData,
-    clientDataJSON,
-    challengeLocation,
-    responseTypeLocation,
-    r,
-    s,
-  }
-
-  const encodedWebAuthnSig = encodeAbiParameters(
-    getAbiItem({
-      abi: sendAccountAbi,
-      name: 'signatureStruct',
-    }).inputs,
-    [webauthnSig]
-  )
-  assert(isHex(encodedWebAuthnSig), 'Invalid encodedWebAuthnSig')
-
-  // @todo: verify signature with user's identifier to ensure it's the correct passkey
-  // const encodedWebAuthnSigBytes = hexToBytes(encodedWebAuthnSig)
-  // const newEncodedWebAuthnSigBytes = new Uint8Array(encodedWebAuthnSigBytes.length + 1)
-  // newEncodedWebAuthnSigBytes[0] = keySlot
-  // newEncodedWebAuthnSigBytes.set(encodedWebAuthnSigBytes, 1)
-  // const verified = await verifySignature(challenge, bytesToHex(newEncodedWebAuthnSigBytes), [
-  //   '0x5BCEE51E9210DAF159CC89BCFDA7FF0AE8AF0881A67D91082503BA90106878D0',
-  //   '0x02CC25B94834CD8214E579356848281F286DD9AED9E5E4D7DD58353990ADD661',
-  // ])
-  // console.log('verified', verified)
-
-  return {
-    keySlot: signResult.keySlot,
-    accountName: signResult.accountName,
-    encodedWebAuthnSig,
   }
 }
